pyproject.toml
[project]
name = "llmorigen"
version = "0.1.0"
description = ""
authors = [
    {name = "Ahmad", email = "ranaahmadintisar@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "blinker==1.9.0",
    "click==8.1.8",
    "flask==3.1.0",
    "itsdangerous==2.2.0",
    "jinja2==3.1.5",
    "markupsafe==3.0.2",
    "werkzeug==3.1.3",
    "flask-sqlalchemy>=3.1.1,<4.0.0",
    
    # Additional dependencies
    "aiofiles==23.2.1",
    "aiohttp==3.9.1",
    "aiosignal==1.3.1",
    "anyio==3.7.1",
    "attrs==23.2.0",
    "bcrypt==4.1.2",
    "beautifulsoup4==4.12.2",
    "bs4==0.0.1",
    "certifi==2023.11.17",
    "cffi==1.16.0",
    "charset-normalizer==3.3.2",
    "chromadb==0.4.22",
    "colorama==0.4.6",
    "duckduckgo-search==4.1.1",
    "fastapi==0.100.1",
    "google-auth==2.26.1",
    "googleapis-common-protos==1.62.0",
    "gradio-client==0.8.0",
    "greenlet==3.0.3",
    "grpcio==1.60.0",
    "h11==0.14.0",
    "httpx==0.24.1",
    "huggingface-hub==0.20.2",
    "idna==3.6",
    "langchain==0.0.354",
    "langchain-core==0.1.6",
    "lxml==5.0.0",
    "matplotlib==3.8.2",
    "mpmath==1.3.0",
    "numpy==1.26.3",
    "onnxruntime==1.16.3",
    "openai==0.28.0",
    "orjson==3.9.10",
    "packaging==23.2",
    "pandas==2.1.4",
    "pillow==10.2.0",
    "protobuf==4.25.1",
    "pydantic==2.5.1",
    "pydantic-core==2.14.3",
    "pydub==0.25.1",
    "pypdf==3.17.4",
    "pyyaml==6.0.1",
    "requests==2.31.0",
    "rich==13.7.0",
    "sqlalchemy==2.0.25",
    "tenacity==8.2.3",
    "tiktoken==0.5.2",
    "tokenizers==0.15.0",
    "typer==0.9.0",
    "typing-extensions==4.9.0",
    "uvicorn==0.23.2",
    "watchfiles==0.20.0",
    "websocket-client==1.7.0",
    "websockets==11.0.3"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
